requirements.txt
python-dotenv
OpenAI
